Refactored convertYCbCrToRGB_Combined

+++ sos_decoder.hpp
@@ -74,10 +74,10 @@ public:
              boost::numeric::ublas::matrix<int16_t> const &cb,
              boost::numeric::ublas::matrix<int16_t> const &cr) ->
     std::tuple<
-        boost::numeric::ublas::matrix<uint16_t>,
-        boost::numeric::ublas::matrix<uint16_t>,
-        boost::numeric::ublas::matrix<uint16_t>
-  >;
+        boost::numeric::ublas::matrix<uint8_t>,
+        boost::numeric::ublas::matrix<uint8_t>,
+        boost::numeric::ublas::matrix<uint8_t>
+    >;
 
 private:
   void InvokeImpl(std::istream &Stream, Context &Ctx) final;
